+++ workspaces/homepage/plugins/homepage/src/alpha/index.ts
@@ -18,6 +18,7 @@ import { TranslationBlueprint } from '@backstage/plugin-app-react';
 import { createFrontendModule } from '@backstage/frontend-plugin-api';
 import {
   catalogStarredWidget,
+  disableRandomJoke,
   disableToolkit,
   entitySectionWidget,
   featuredDocsCardWidget,
@@ -58,6 +59,7 @@ export const homePageModule = createFrontendModule({
     RecentlyVisitedWidget,
     catalogStarredWidget,
     disableToolkit,
+    disableRandomJoke,
   ],
 });