+++ workspaces/homepage/plugins/homepage-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @red-hat-developer-hub/backstage-plugin-homepage-backend
 
+## 0.4.0
+
+### Minor Changes
+
+- a6fced5: Backstage version bump to v1.52.1
+
+### Patch Changes
+
+- Updated dependencies [a6fced5]
+  - @red-hat-developer-hub/backstage-plugin-homepage-common@0.4.0
+
 ## 0.3.0
 
 ### Minor Changes